Add cursor pagination for sources.
